Refactor get-user script
Move functions of get-user script into helpers and separate module.

This will make it easier to reuse this functionality in the future/use
this functionality from the api shell.
